+++ jest-config/jest-integration-local.config.js
@@ -1,0 +1,9 @@
+// eslint-disable-next-line @typescript-eslint/no-var-requires
+const configIntegration = require("./jest-integration.config");
+/** @type any */
+module.exports = {
+  ...configIntegration,
+  testPathIgnorePatterns: configIntegration.testPathIgnorePatterns.filter(
+    (e) => e !== "intellisense"
+  ),
+};