+++ src/plugins/intellisense/index.tsx
@@ -6,6 +6,7 @@ import {
   getPartialFunctionCall,
 } from "./latex-parsing";
 import { IntellisenseState } from "./state";
+import { pendingIntellisenseTimeouts, setIntellisenseTimeout } from "./utils";
 import { JumpToDefinitionMenuInfo, View } from "./view";
 import { DCGView, MountedComponent, unmountFromNode } from "DCGView";
 import { MathQuillField, MathQuillView } from "components";
@@ -65,7 +66,7 @@ export function getExpressionLatex(id: string): string | undefined {
 
 export default class Intellisense extends PluginController {
   static id = "intellisense" as const;
-  static enabledByDefault = true;
+  static enabledByDefault = false;
   static descriptionLearnMore =
     "https://github.com/DesModder/DesModder/tree/main/src/plugins/intellisense/docs/README.md";
 
@@ -101,6 +102,22 @@ export default class Intellisense extends PluginController {
   jumpToDefIndex: number = 0;
 
   specialIdentifierNames: string[] = [];
+
+  async waitForCurrentIntellisenseTimeoutsToFinish() {
+    await new Promise<void>((resolve) => {
+      const currentTimeouts = Array.from(pendingIntellisenseTimeouts.entries());
+      const interval = setInterval(() => {
+        // only continue if all timeouts have been finished
+        for (const timeout of currentTimeouts) {
+          if (pendingIntellisenseTimeouts.has(timeout)) return;
+        }
+
+        // resolve the promise when timeouts have finished
+        clearInterval(interval);
+        resolve();
+      });
+    });
+  }
 
   // recalculate the intellisense
   updateIntellisense() {
@@ -319,7 +336,7 @@ export default class Intellisense extends PluginController {
   }
 
   focusInHandler = () => {
-    setTimeout(() => {
+    setIntellisenseTimeout(() => {
       if (Calc.focusedMathQuill && this.specialIdentifierNames.length === 0) {
         this.specialIdentifierNames = [
           ...Object.keys(Calc.focusedMathQuill.mq.__options.autoOperatorNames),
@@ -375,7 +392,7 @@ export default class Intellisense extends PluginController {
 
                 // need a delay so that Enter key doesn't immediately close
                 // the jump2def window
-                setTimeout(() => {
+                setIntellisenseTimeout(() => {
                   self.jumpToDefinition(str);
                 });
               }
@@ -418,6 +435,17 @@ export default class Intellisense extends PluginController {
     });
   };
 
+  // allows mathquill inputs that only allow arithmetic to selectively disable intellisense
+  isActiveElementValidForIntellisense() {
+    return (
+      document.activeElement
+        ?.closest(
+          ".yes-intellisense, .no-intellisense, .dcg-settings-view-container"
+        )
+        ?.classList.contains("yes-intellisense") ?? true
+    );
+  }
+
   keyDownHandler = (e: KeyboardEvent) => {
     this.saveCursorState();
 
@@ -436,7 +464,12 @@ export default class Intellisense extends PluginController {
 
     // if a non arrow key is pressed in an expression,
     // we enable the intellisense window
-    if (!e.key.startsWith("Arrow") && e.key !== "Enter" && e.key !== "Escape") {
+    if (
+      !e.key.startsWith("Arrow") &&
+      e.key !== "Enter" &&
+      e.key !== "Escape" &&
+      this.isActiveElementValidForIntellisense()
+    ) {
       this.canHaveIntellisense = true;
     }
 
@@ -644,7 +677,7 @@ export default class Intellisense extends PluginController {
 
     this.dispatcher = Calc.controller.dispatcher.register((e) => {
       if (e.type === "set-focus-location" || e.type === "set-none-selected") {
-        setTimeout(() => {
+        setIntellisenseTimeout(() => {
           if (!Calc.focusedMathQuill) {
             this.canHaveIntellisense = false;
             this.view?.update();